(PC-8100): Extract synchronize stock in specific api action
