+++ src/pcapi/local_providers/provider_api/synchronize_provider_api.py
@@ -1,28 +1,18 @@
 from datetime import datetime
 import logging
 import time
+from typing import Counter
+from typing import Dict
 from typing import Generator
-from typing import Union
+from typing import List
 
-from flask import current_app as app
 from sqlalchemy.sql.sqltypes import DateTime
 
-from pcapi.connectors import redis
-from pcapi.core.offers.models import Offer
-from pcapi.core.offers.models import Stock
-from pcapi.core.offers.repository import get_offers_map_by_id_at_providers
-from pcapi.core.offers.repository import get_products_map_by_id_at_providers
-from pcapi.core.offers.repository import get_stocks_by_id_at_providers
+from pcapi.core.providers.api import synchronize_stocks
 from pcapi.core.providers.models import Provider
 from pcapi.core.providers.models import VenueProvider
 from pcapi.infrastructure.repository.stock_provider.provider_api import ProviderAPI
-from pcapi.models import Product
-from pcapi.models import Venue
-from pcapi.models.db import db
-from pcapi.models.feature import FeatureToggle
-from pcapi.repository import feature_queries
 from pcapi.repository import repository
-from pcapi.validation.models.entity_validator import validate
 
 
 logger = logging.getLogger(__name__)
@@ -42,50 +32,13 @@ def synchronize_venue_provider(venue_provider: VenueProvider) -> None:
     logger.info("Starting synchronization of venue=%s provider=%s", venue.id, provider.name)
     provider_api = provider.getProviderAPI()
 
-    stats = {"new_offers": 0, "new_stocks": 0, "updated_stocks": 0}
+    stats = Counter()
     for raw_stocks in _get_stocks_by_batch(
         venue_provider.venueIdAtOfferProvider, provider_api, venue_provider.lastSyncDate
     ):
         stock_details = _build_stock_details_from_raw_stocks(raw_stocks, venue_provider.venueIdAtOfferProvider)
-
-        products_provider_references = [stock_detail["products_provider_reference"] for stock_detail in stock_details]
-        products_by_provider_reference = get_products_map_by_id_at_providers(products_provider_references)
-
-        stock_details = [
-            stock for stock in stock_details if stock["products_provider_reference"] in products_by_provider_reference
-        ]
-
-        offers_provider_references = [stock_detail["offers_provider_reference"] for stock_detail in stock_details]
-        offers_by_provider_reference = get_offers_map_by_id_at_providers(offers_provider_references)
-
-        new_offers = _build_new_offers_from_stock_details(
-            stock_details, offers_by_provider_reference, products_by_provider_reference, venue_provider
-        )
-        new_offers_references = [new_offer.idAtProviders for new_offer in new_offers]
-
-        db.session.bulk_save_objects(new_offers)
-        stats["new_offers"] += len(new_offers)
-
-        new_offers_by_provider_reference = get_offers_map_by_id_at_providers(new_offers_references)
-        offers_by_provider_reference = {**offers_by_provider_reference, **new_offers_by_provider_reference}
-
-        stocks_provider_references = [stock["stocks_provider_reference"] for stock in stock_details]
-        stocks_by_provider_reference = get_stocks_by_id_at_providers(stocks_provider_references)
-        update_stock_mapping, new_stocks, offer_ids = _get_stocks_to_upsert(
-            stock_details,
-            stocks_by_provider_reference,
-            offers_by_provider_reference,
-            products_by_provider_reference,
-        )
-
-        db.session.bulk_save_objects(new_stocks)
-        db.session.bulk_update_mappings(Stock, update_stock_mapping)
-        stats["new_stocks"] += len(new_stocks)
-        stats["updated_stocks"] += len(update_stock_mapping)
-
-        db.session.commit()
-
-        _reindex_offers(offer_ids)
+        operations = synchronize_stocks(stock_details, venue, provider_id=provider.id)
+        stats += Counter(operations)
 
     venue_provider.lastSyncDate = start_sync_date
     repository.save(venue_provider)
@@ -132,117 +85,3 @@ def _build_stock_details_from_raw_stocks(raw_stocks: list[dict], venue_siret: st
         }
 
     return list(stock_details.values())
-
-
-def _build_new_offers_from_stock_details(
-    stock_details: list,
-    existing_offers_by_provider_reference: dict[str, int],
-    products_by_provider_reference: dict[str, Product],
-    venue_provider: VenueProvider,
-) -> list[Offer]:
-    new_offers = []
-    for stock_detail in stock_details:
-        if stock_detail["offers_provider_reference"] in existing_offers_by_provider_reference:
-            continue
-
-        product = products_by_provider_reference[stock_detail["products_provider_reference"]]
-        offer = _build_new_offer(
-            venue_provider.venue,
-            product,
-            id_at_providers=stock_detail["offers_provider_reference"],
-            provider_id=venue_provider.providerId,
-        )
-
-        if not _validate_stock_or_offer(offer):
-            continue
-
-        new_offers.append(offer)
-
-    return new_offers
-
-
-def _get_stocks_to_upsert(
-    stock_details: list[dict],
-    stocks_by_provider_reference: dict[str, dict],
-    offers_by_provider_reference: dict[str, int],
-    products_by_provider_reference: dict[str, Product],
-) -> tuple[list[dict], list[Stock], set[int]]:
-    update_stock_mapping = []
-    new_stocks = []
-    offer_ids = set()
-
-    for stock_detail in stock_details:
-        stock_provider_reference = stock_detail["stocks_provider_reference"]
-        product = products_by_provider_reference[stock_detail["products_provider_reference"]]
-        book_price = float(product.extraData["prix_livre"])
-        if stock_provider_reference in stocks_by_provider_reference:
-            stock = stocks_by_provider_reference[stock_provider_reference]
-
-            update_stock_mapping.append(
-                {
-                    "id": stock["id"],
-                    "quantity": stock_detail["available_quantity"] + stock["booking_quantity"],
-                    "rawProviderQuantity": stock_detail["available_quantity"],
-                    "price": book_price,
-                }
-            )
-            offer_ids.add(offers_by_provider_reference[stock_detail["offers_provider_reference"]])
-
-        else:
-            stock = _build_stock_from_stock_detail(
-                stock_detail,
-                offers_by_provider_reference[stock_detail["offers_provider_reference"]],
-                book_price,
-            )
-            if not _validate_stock_or_offer(stock):
-                continue
-
-            new_stocks.append(stock)
-            offer_ids.add(stock.offerId)
-
-    return update_stock_mapping, new_stocks, offer_ids
-
-
-def _build_stock_from_stock_detail(stock_detail: dict, offers_id: int, price: float) -> Stock:
-    return Stock(
-        quantity=stock_detail["available_quantity"],
-        rawProviderQuantity=stock_detail["available_quantity"],
-        bookingLimitDatetime=None,
-        offerId=offers_id,
-        price=price,
-        dateModified=datetime.now(),
-        idAtProviders=stock_detail["stocks_provider_reference"],
-    )
-
-
-def _validate_stock_or_offer(model: Union[Offer, Stock]) -> bool:
-    model_api_errors = validate(model)
-    if model_api_errors.errors.keys():
-        logger.exception(
-            "[SYNC] errors while trying to add stock or offer with ref %s: %s",
-            model.idAtProviders,
-            model_api_errors.errors,
-        )
-        return False
-
-    return True
-
-
-def _build_new_offer(venue: Venue, product: Product, id_at_providers: str, provider_id: str) -> Offer:
-    return Offer(
-        bookingEmail=venue.bookingEmail,
-        description=product.description,
-        extraData=product.extraData,
-        idAtProviders=id_at_providers,
-        lastProviderId=provider_id,
-        name=product.name,
-        productId=product.id,
-        venueId=venue.id,
-        type=product.type,
-    )
-
-
-def _reindex_offers(offer_ids: set[int]) -> None:
-    if feature_queries.is_active(FeatureToggle.SYNCHRONIZE_ALGOLIA):
-        for offer_id in offer_ids:
-            redis.add_offer_id(client=app.redis_client, offer_id=offer_id)
